+++ libsimple-chess/ui/SymbolicBoard.hpp
@@ -1,7 +1,7 @@
 /*
  -------------------------------------------------------------------------------
     This file is part of simple-chess.
-    Copyright (C) 2016, 2021, 2022  Dirk Stolle
+    Copyright (C) 2022  Dirk Stolle
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -18,27 +18,25 @@
  -------------------------------------------------------------------------------
 */
 
-#include "Console.hpp"
-#include <iostream>
-#include "letters.hpp"
+#ifndef SIMPLE_CHESS_UI_SYMBOLIC_BOARD_HPP
+#define SIMPLE_CHESS_UI_SYMBOLIC_BOARD_HPP
+
+#include "../data/Board.hpp"
 
 namespace simplechess::ui
 {
 
-void Console::showBoard(const Board & board)
+/** Shows a chess board on the console using Unicode symbols for pieces. */
+class SymbolicBoard
 {
-  const std::string rowSeparator = "+---+---+---+---+---+---+---+---+";
-  std::cout << rowSeparator << "\n";
-  for (int r = 8; r >= 1; r--)
-  {
-    std::cout << "| ";
-    for (char c = 'a'; c <= 'h'; ++c)
-    {
-      const Piece & piece = board.element(toField(c, r));
-      std::cout << letter(piece) << " | ";
-    } // for c
-    std::cout << "\n" << rowSeparator << "\n";
-  } // for r
-}
+  public:
+    /** \brief Prints the board to standard output.
+     *
+     * \param board   the current chess board
+     */
+    static void showBoard(const Board& board);
+};
 
 } // namespace
+
+#endif // SIMPLE_CHESS_UI_SYMBOLIC_BOARD_HPP